re-inline type feedback
