+++ rir/src/interpreter/interp.cpp
@@ -9,6 +9,7 @@
 #include "compiler/translations/rir_2_pir/rir_2_pir_compiler.h"
 #include "event_counters.h"
 #include "ir/Deoptimization.h"
+#include "runtime/TypeFeedback_inl.h"
 #include "safe_force.h"
 #include "utils/Pool.h"
