+++ rir/src/runtime/TypeFeedback.cpp
@@ -6,25 +6,6 @@
 
 namespace rir {
 
-ObservedType::ObservedType(SEXP s)
-    : sexptype((uint8_t)TYPEOF(s)), scalar(IS_SCALAR(s, TYPEOF(s))),
-      object(OBJECT(s)), attribs(ATTRIB(s) != R_NilValue) {}
-
-void ObservedCallees::record(Code* caller, SEXP callee) {
-    if (taken < CounterOverflow)
-        taken++;
-    if (numTargets < MaxTargets) {
-        int i = 0;
-        for (; i < numTargets; ++i)
-            if (caller->getExtraPoolEntry(targets[i]) == callee)
-                break;
-        if (i == numTargets) {
-            auto idx = caller->addExtraPoolEntry(callee);
-            targets[numTargets++] = idx;
-        }
-    }
-}
-
 SEXP ObservedCallees::getTarget(const Code* code, size_t pos) const {
     assert(pos < numTargets);
     return code->getExtraPoolEntry(targets[pos]);
